Addressed comments about directories, etc.
