+++ examples/Travel/AddHotelAd.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright 2023 Google LLC
+ * Copyright 2018 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-namespace Google\Ads\GoogleAds\Examples\HotelAndTravel;
+namespace Google\Ads\GoogleAds\Examples\Travel;
 
 require __DIR__ . '/../../vendor/autoload.php';
 
@@ -24,16 +24,15 @@ use GetOpt\GetOpt;
 use Google\Ads\GoogleAds\Examples\Utils\ArgumentNames;
 use Google\Ads\GoogleAds\Examples\Utils\ArgumentParser;
 use Google\Ads\GoogleAds\Examples\Utils\Helper;
-use Google\Ads\GoogleAds\Lib\OAuth2TokenBuilder;
 use Google\Ads\GoogleAds\Lib\V13\GoogleAdsClient;
 use Google\Ads\GoogleAds\Lib\V13\GoogleAdsClientBuilder;
 use Google\Ads\GoogleAds\Lib\V13\GoogleAdsException;
-use Google\Ads\GoogleAds\V13\Common\MaximizeConversionValue;
-use Google\Ads\GoogleAds\V13\Common\TravelAdInfo;
+use Google\Ads\GoogleAds\Lib\OAuth2TokenBuilder;
+use Google\Ads\GoogleAds\V13\Common\HotelAdInfo;
+use Google\Ads\GoogleAds\V13\Common\PercentCpc;
 use Google\Ads\GoogleAds\V13\Enums\AdGroupAdStatusEnum\AdGroupAdStatus;
 use Google\Ads\GoogleAds\V13\Enums\AdGroupStatusEnum\AdGroupStatus;
 use Google\Ads\GoogleAds\V13\Enums\AdGroupTypeEnum\AdGroupType;
-use Google\Ads\GoogleAds\V13\Enums\AdvertisingChannelSubTypeEnum\AdvertisingChannelSubType;
 use Google\Ads\GoogleAds\V13\Enums\AdvertisingChannelTypeEnum\AdvertisingChannelType;
 use Google\Ads\GoogleAds\V13\Enums\BudgetDeliveryMethodEnum\BudgetDeliveryMethod;
 use Google\Ads\GoogleAds\V13\Enums\CampaignStatusEnum\CampaignStatus;
@@ -42,8 +41,8 @@ use Google\Ads\GoogleAds\V13\Resources\Ad;
 use Google\Ads\GoogleAds\V13\Resources\AdGroup;
 use Google\Ads\GoogleAds\V13\Resources\AdGroupAd;
 use Google\Ads\GoogleAds\V13\Resources\Campaign;
+use Google\Ads\GoogleAds\V13\Resources\Campaign\HotelSettingInfo;
 use Google\Ads\GoogleAds\V13\Resources\Campaign\NetworkSettings;
-use Google\Ads\GoogleAds\V13\Resources\Campaign\TravelCampaignSettings;
 use Google\Ads\GoogleAds\V13\Resources\CampaignBudget;
 use Google\Ads\GoogleAds\V13\Services\AdGroupAdOperation;
 use Google\Ads\GoogleAds\V13\Services\AdGroupOperation;
@@ -52,16 +51,23 @@ use Google\Ads\GoogleAds\V13\Services\CampaignOperation;
 use Google\ApiCore\ApiException;
 
 /**
- * This example creates a Things to do campaign, an ad group and a Things to do ad.
+ * This example creates a hotel campaign, a hotel ad group and hotel ad group ad.
  *
- * <p> Prerequisite: You need to have an access to the Things to Do Center. The integration
- * instructions can be found at: https://support.google.com/google-ads/answer/13387362.
+ * <p> Prerequisite: You need to have an access to the Hotel Ads Center, which can be granted during
+ * integration with Google Hotels. The integration instructions can be found at:
+ * https://support.google.com/hotelprices/answer/6101897.
  */
-class AddThingsToDoAd
+class AddHotelAd
 {
     private const CUSTOMER_ID = 'INSERT_CUSTOMER_ID_HERE';
-    // Specify your Things to Do Center account ID below.
-    private const THINGS_TO_DO_CENTER_ACCOUNT_ID = 'INSERT_THINGS_TO_DO_CENTER_ACCOUNT_ID_HERE';
+    // Specify your Hotels account ID below. You can see how to find the account ID in the Hotel
+    // Ads Center at: https://support.google.com/hotelprices/answer/6399770.
+    // This ID is the same account ID that you use in API requests to the Travel Partner APIs
+    // (https://developers.google.com/hotels/hotel-ads/api-reference/).
+    private const HOTEL_CENTER_ACCOUNT_ID = 'INSERT_HOTEL_CENTER_ACCOUNT_ID_HERE';
+    // Specify maximum bid limit that can be set when creating a campaign using the Percent CPC
+    // bidding strategy.
+    private const CPC_BID_CEILING_MICRO_AMOUNT = 20000000;
 
     public static function main()
     {
@@ -69,7 +75,8 @@ class AddThingsToDoAd
         // into the constants above.
         $options = (new ArgumentParser())->parseCommandArguments([
             ArgumentNames::CUSTOMER_ID => GetOpt::REQUIRED_ARGUMENT,
-            ArgumentNames::THINGS_TO_DO_CENTER_ACCOUNT_ID => GetOpt::REQUIRED_ARGUMENT
+            ArgumentNames::HOTEL_CENTER_ACCOUNT_ID => GetOpt::REQUIRED_ARGUMENT,
+            ArgumentNames::CPC_BID_CEILING_MICRO_AMOUNT => GetOpt::OPTIONAL_ARGUMENT
         ]);
 
         // Generate a refreshable OAuth2 credential for authentication.
@@ -85,8 +92,9 @@ class AddThingsToDoAd
             self::runExample(
                 $googleAdsClient,
                 $options[ArgumentNames::CUSTOMER_ID] ?: self::CUSTOMER_ID,
-                $options[ArgumentNames::THINGS_TO_DO_CENTER_ACCOUNT_ID]
-                    ?: self::THINGS_TO_DO_CENTER_ACCOUNT_ID
+                $options[ArgumentNames::HOTEL_CENTER_ACCOUNT_ID] ?: self::HOTEL_CENTER_ACCOUNT_ID,
+                $options[ArgumentNames::CPC_BID_CEILING_MICRO_AMOUNT]
+                    ?: self::CPC_BID_CEILING_MICRO_AMOUNT
             );
         } catch (GoogleAdsException $googleAdsException) {
             printf(
@@ -120,31 +128,34 @@ class AddThingsToDoAd
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
-     * @param int $thingsToDoCenterAccountId the Things to Do Center account ID
+     * @param int $hotelCenterAccountId the Hotel Center account ID
+     * @param int $cpcBidCeilingMicroAmount the CPC bid ceiling micro amount
      */
     public static function runExample(
         GoogleAdsClient $googleAdsClient,
         int $customerId,
-        int $thingsToDoCenterAccountId
+        int $hotelCenterAccountId,
+        int $cpcBidCeilingMicroAmount
     ) {
         // Creates a budget to be used by the campaign that will be created below.
         $budgetResourceName = self::addCampaignBudget($googleAdsClient, $customerId);
-        // Creates a Things to do campaign.
-        $campaignResourceName = self::addThingsToDoCampaign(
+        // Creates a hotel campaign.
+        $campaignResourceName = self::addHotelCampaign(
             $googleAdsClient,
             $customerId,
             $budgetResourceName,
-            $thingsToDoCenterAccountId
+            $hotelCenterAccountId,
+            $cpcBidCeilingMicroAmount
         );
-        // Creates an ad group.
+        // Creates a hotel ad group.
         $adGroupResourceName =
-            self::addAdGroup($googleAdsClient, $customerId, $campaignResourceName);
-        // Creates an ad group ad.
-        self::addAdGroupAd($googleAdsClient, $customerId, $adGroupResourceName);
+            self::addHotelAdGroup($googleAdsClient, $customerId, $campaignResourceName);
+        // Creates a hotel ad group ad.
+        self::addHotelAdGroupAd($googleAdsClient, $customerId, $adGroupResourceName);
     }
 
     /**
-     * Creates a new campaign budget in the specified customer account.
+     * Creates a new campaign budget in the specified client account.
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
@@ -185,45 +196,49 @@ class AddThingsToDoAd
     }
 
     /**
-     * Creates a new Things to do campaign in the specified customer account.
+     * Creates a new hotel campaign in the specified client account.
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
      * @param string $budgetResourceName the resource name of budget for a new campaign
-     * @param int $thingsToDoCenterAccountId the Things to Do Center account ID
+     * @param int $hotelCenterAccountId the Hotel Center account ID
+     * @param int $cpcBidCeilingMicroAmount the CPC bid ceiling micro amount
      * @return string the resource name of the newly created campaign
      */
-    // [START add_things_to_do_ad]
-    private static function addThingsToDoCampaign(
+    // [START add_hotel_ad]
+    private static function addHotelCampaign(
         GoogleAdsClient $googleAdsClient,
         int $customerId,
         string $budgetResourceName,
-        int $thingsToDoCenterAccountId
+        int $hotelCenterAccountId,
+        int $cpcBidCeilingMicroAmount
     ) {
-        // [START add_things_to_do_ad_1]
+        // [START add_hotel_ad_1]
         // Creates a campaign.
         $campaign = new Campaign([
             'name' => 'Interplanetary Cruise Campaign #' . Helper::getPrintableDatetime(),
-            // Configures settings related to Things to do campaigns including advertising channel
-            // type, advertising channel sub type and travel campaign settings.
-            'advertising_channel_type' => AdvertisingChannelType::TRAVEL,
-            'advertising_channel_sub_type' => AdvertisingChannelSubType::TRAVEL_ACTIVITIES,
-            'travel_campaign_settings'
-                => new TravelCampaignSettings(['travel_account_id' => $thingsToDoCenterAccountId]),
+            // Configures settings related to hotel campaigns including advertising channel type
+            // and hotel setting info.
+            'advertising_channel_type' => AdvertisingChannelType::HOTEL,
+            'hotel_setting' => new HotelSettingInfo(['hotel_center_id' => $hotelCenterAccountId]),
             // Recommendation: Set the campaign to PAUSED when creating it to prevent
             // the ads from immediately serving. Set to ENABLED once you've added
             // targeting and the ads are ready to serve.
             'status' => CampaignStatus::PAUSED,
-            // Sets the bidding strategy to MaximizeConversionValue. Only this type can be used
-            // for Things to do campaigns.
-            'maximize_conversion_value' => new MaximizeConversionValue(),
+            // Sets the bidding strategy to PercentCpc. Only Manual CPC and Percent CPC can be used
+            // for hotel campaigns.
+            'percent_cpc' => new PercentCpc([
+                'cpc_bid_ceiling_micros' => $cpcBidCeilingMicroAmount
+            ]),
             // Sets the budget.
             'campaign_budget' => $budgetResourceName,
             // Configures the campaign network options. Only Google Search is allowed for
-            // Things to do campaigns.
-            'network_settings' => new NetworkSettings(['target_google_search' => true])
+            // hotel campaigns.
+            'network_settings' => new NetworkSettings([
+                'target_google_search' => true,
+            ]),
         ]);
-        // [END add_things_to_do_ad_1]
+        // [END add_hotel_ad_1]
 
         // Creates a campaign operation.
         $campaignOperation = new CampaignOperation();
@@ -236,17 +251,17 @@ class AddThingsToDoAd
         /** @var Campaign $addedCampaign */
         $addedCampaign = $response->getResults()[0];
         printf(
-            "Added a Things to do campaign with resource name '%s'.%s",
+            "Added a hotel campaign with resource name '%s'.%s",
             $addedCampaign->getResourceName(),
             PHP_EOL
         );
 
         return $addedCampaign->getResourceName();
     }
-    // [END add_things_to_do_ad]
+    // [END add_hotel_ad]
 
     /**
-     * Creates a new ad group in the specified Things to do campaign.
+     * Creates a new hotel ad group in the specified campaign.
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
@@ -254,8 +269,8 @@ class AddThingsToDoAd
      *     belong to
      * @return string the resource name of the newly created ad group
      */
-    // [START add_things_to_do_ad_2]
-    private static function addAdGroup(
+    // [START add_hotel_ad_2]
+    private static function addHotelAdGroup(
         GoogleAdsClient $googleAdsClient,
         int $customerId,
         string $campaignResourceName
@@ -265,8 +280,9 @@ class AddThingsToDoAd
             'name' => 'Earth to Mars Cruise #' . Helper::getPrintableDatetime(),
             // Sets the campaign.
             'campaign' => $campaignResourceName,
-            // Sets the ad group type to TRAVEL_ADS. This cannot be set to other types.
-            'type' => AdGroupType::TRAVEL_ADS,
+            // Sets the ad group type to HOTEL_ADS.
+            // This cannot be set to other types.
+            'type' => AdGroupType::HOTEL_ADS,
             'cpc_bid_micros' => 10000000,
             'status' => AdGroupStatus::ENABLED,
         ]);
@@ -282,34 +298,40 @@ class AddThingsToDoAd
         /** @var AdGroup $addedAdGroup */
         $addedAdGroup = $response->getResults()[0];
         printf(
-            "Added an ad group with resource name '%s'.%s",
+            "Added a hotel ad group with resource name '%s'.%s",
             $addedAdGroup->getResourceName(),
             PHP_EOL
         );
 
         return $addedAdGroup->getResourceName();
     }
-    // [END add_things_to_do_ad_2]
+    // [END add_hotel_ad_2]
 
     /**
-     * Creates a new ad group ad in the specified ad group.
+     * Creates a new hotel ad group ad in the specified ad group.
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
      * @param string $adGroupResourceName the resource name of ad group that a new ad group ad will
      *     belong to
      */
-    // [START add_things_to_do_ad_3]
-    private static function addAdGroupAd(
+    // [START add_hotel_ad_3]
+    private static function addHotelAdGroupAd(
         GoogleAdsClient $googleAdsClient,
         int $customerId,
         string $adGroupResourceName
     ) {
-        // Creates a new ad group ad and set a travel ad info.
+        // Creates a new hotel ad.
+        $ad = new Ad([
+            'hotel_ad' => new HotelAdInfo(),
+        ]);
+
+        // Creates a new ad group ad and sets the hotel ad to it.
         $adGroupAd = new AdGroupAd([
-            'ad' => new Ad(['travel_ad' => new TravelAdInfo()]),
-            // Set the ad group ad to enabled. Setting this to paused will cause an error for Things
-            // to do campaigns. Pausing should happen at either the ad group or campaign level.
+            'ad' => $ad,
+            // Set the ad group ad to enabled.  Setting this to paused will cause an error
+            // for hotel campaigns.  For hotels pausing should happen at either the ad group or
+            // campaign level.
             'status' => AdGroupAdStatus::ENABLED,
             // Sets the ad group.
             'ad_group' => $adGroupResourceName
@@ -326,12 +348,12 @@ class AddThingsToDoAd
         /** @var AdGroupAd $addedAdGroupAd */
         $addedAdGroupAd = $response->getResults()[0];
         printf(
-            "Added an ad group ad with resource name '%s'.%s",
+            "Added a hotel ad group ad with resource name '%s'.%s",
             $addedAdGroupAd->getResourceName(),
             PHP_EOL
         );
     }
-    // [END add_things_to_do_ad_3]
+    // [END add_hotel_ad_3]
 }
 
-AddThingsToDoAd::main();
+AddHotelAd::main();
