+++ examples/Travel/AddThingsToDoAd.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright 2018 Google LLC
+ * Copyright 2023 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-namespace Google\Ads\GoogleAds\Examples\HotelAndTravel;
+namespace Google\Ads\GoogleAds\Examples\Travel;
 
 require __DIR__ . '/../../vendor/autoload.php';
 
@@ -24,15 +24,16 @@ use GetOpt\GetOpt;
 use Google\Ads\GoogleAds\Examples\Utils\ArgumentNames;
 use Google\Ads\GoogleAds\Examples\Utils\ArgumentParser;
 use Google\Ads\GoogleAds\Examples\Utils\Helper;
+use Google\Ads\GoogleAds\Lib\OAuth2TokenBuilder;
 use Google\Ads\GoogleAds\Lib\V13\GoogleAdsClient;
 use Google\Ads\GoogleAds\Lib\V13\GoogleAdsClientBuilder;
 use Google\Ads\GoogleAds\Lib\V13\GoogleAdsException;
-use Google\Ads\GoogleAds\Lib\OAuth2TokenBuilder;
-use Google\Ads\GoogleAds\V13\Common\HotelAdInfo;
-use Google\Ads\GoogleAds\V13\Common\PercentCpc;
+use Google\Ads\GoogleAds\V13\Common\MaximizeConversionValue;
+use Google\Ads\GoogleAds\V13\Common\TravelAdInfo;
 use Google\Ads\GoogleAds\V13\Enums\AdGroupAdStatusEnum\AdGroupAdStatus;
 use Google\Ads\GoogleAds\V13\Enums\AdGroupStatusEnum\AdGroupStatus;
 use Google\Ads\GoogleAds\V13\Enums\AdGroupTypeEnum\AdGroupType;
+use Google\Ads\GoogleAds\V13\Enums\AdvertisingChannelSubTypeEnum\AdvertisingChannelSubType;
 use Google\Ads\GoogleAds\V13\Enums\AdvertisingChannelTypeEnum\AdvertisingChannelType;
 use Google\Ads\GoogleAds\V13\Enums\BudgetDeliveryMethodEnum\BudgetDeliveryMethod;
 use Google\Ads\GoogleAds\V13\Enums\CampaignStatusEnum\CampaignStatus;
@@ -41,8 +42,8 @@ use Google\Ads\GoogleAds\V13\Resources\Ad;
 use Google\Ads\GoogleAds\V13\Resources\AdGroup;
 use Google\Ads\GoogleAds\V13\Resources\AdGroupAd;
 use Google\Ads\GoogleAds\V13\Resources\Campaign;
-use Google\Ads\GoogleAds\V13\Resources\Campaign\HotelSettingInfo;
 use Google\Ads\GoogleAds\V13\Resources\Campaign\NetworkSettings;
+use Google\Ads\GoogleAds\V13\Resources\Campaign\TravelCampaignSettings;
 use Google\Ads\GoogleAds\V13\Resources\CampaignBudget;
 use Google\Ads\GoogleAds\V13\Services\AdGroupAdOperation;
 use Google\Ads\GoogleAds\V13\Services\AdGroupOperation;
@@ -51,23 +52,16 @@ use Google\Ads\GoogleAds\V13\Services\CampaignOperation;
 use Google\ApiCore\ApiException;
 
 /**
- * This example creates a hotel campaign, a hotel ad group and hotel ad group ad.
+ * This example creates a Things to do campaign, an ad group and a Things to do ad.
  *
- * <p> Prerequisite: You need to have an access to the Hotel Ads Center, which can be granted during
- * integration with Google Hotels. The integration instructions can be found at:
- * https://support.google.com/hotelprices/answer/6101897.
+ * <p> Prerequisite: You need to have an access to the Things to Do Center. The integration
+ * instructions can be found at: https://support.google.com/google-ads/answer/13387362.
  */
-class AddHotelAd
+class AddThingsToDoAd
 {
     private const CUSTOMER_ID = 'INSERT_CUSTOMER_ID_HERE';
-    // Specify your Hotels account ID below. You can see how to find the account ID in the Hotel
-    // Ads Center at: https://support.google.com/hotelprices/answer/6399770.
-    // This ID is the same account ID that you use in API requests to the Travel Partner APIs
-    // (https://developers.google.com/hotels/hotel-ads/api-reference/).
-    private const HOTEL_CENTER_ACCOUNT_ID = 'INSERT_HOTEL_CENTER_ACCOUNT_ID_HERE';
-    // Specify maximum bid limit that can be set when creating a campaign using the Percent CPC
-    // bidding strategy.
-    private const CPC_BID_CEILING_MICRO_AMOUNT = 20000000;
+    // Specify your Things to Do Center account ID below.
+    private const THINGS_TO_DO_CENTER_ACCOUNT_ID = 'INSERT_THINGS_TO_DO_CENTER_ACCOUNT_ID_HERE';
 
     public static function main()
     {
@@ -75,8 +69,7 @@ class AddHotelAd
         // into the constants above.
         $options = (new ArgumentParser())->parseCommandArguments([
             ArgumentNames::CUSTOMER_ID => GetOpt::REQUIRED_ARGUMENT,
-            ArgumentNames::HOTEL_CENTER_ACCOUNT_ID => GetOpt::REQUIRED_ARGUMENT,
-            ArgumentNames::CPC_BID_CEILING_MICRO_AMOUNT => GetOpt::OPTIONAL_ARGUMENT
+            ArgumentNames::THINGS_TO_DO_CENTER_ACCOUNT_ID => GetOpt::REQUIRED_ARGUMENT
         ]);
 
         // Generate a refreshable OAuth2 credential for authentication.
@@ -92,9 +85,8 @@ class AddHotelAd
             self::runExample(
                 $googleAdsClient,
                 $options[ArgumentNames::CUSTOMER_ID] ?: self::CUSTOMER_ID,
-                $options[ArgumentNames::HOTEL_CENTER_ACCOUNT_ID] ?: self::HOTEL_CENTER_ACCOUNT_ID,
-                $options[ArgumentNames::CPC_BID_CEILING_MICRO_AMOUNT]
-                    ?: self::CPC_BID_CEILING_MICRO_AMOUNT
+                $options[ArgumentNames::THINGS_TO_DO_CENTER_ACCOUNT_ID]
+                    ?: self::THINGS_TO_DO_CENTER_ACCOUNT_ID
             );
         } catch (GoogleAdsException $googleAdsException) {
             printf(
@@ -128,34 +120,31 @@ class AddHotelAd
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
-     * @param int $hotelCenterAccountId the Hotel Center account ID
-     * @param int $cpcBidCeilingMicroAmount the CPC bid ceiling micro amount
+     * @param int $thingsToDoCenterAccountId the Things to Do Center account ID
      */
     public static function runExample(
         GoogleAdsClient $googleAdsClient,
         int $customerId,
-        int $hotelCenterAccountId,
-        int $cpcBidCeilingMicroAmount
+        int $thingsToDoCenterAccountId
     ) {
         // Creates a budget to be used by the campaign that will be created below.
         $budgetResourceName = self::addCampaignBudget($googleAdsClient, $customerId);
-        // Creates a hotel campaign.
-        $campaignResourceName = self::addHotelCampaign(
+        // Creates a Things to do campaign.
+        $campaignResourceName = self::addThingsToDoCampaign(
             $googleAdsClient,
             $customerId,
             $budgetResourceName,
-            $hotelCenterAccountId,
-            $cpcBidCeilingMicroAmount
+            $thingsToDoCenterAccountId
         );
-        // Creates a hotel ad group.
+        // Creates an ad group.
         $adGroupResourceName =
-            self::addHotelAdGroup($googleAdsClient, $customerId, $campaignResourceName);
-        // Creates a hotel ad group ad.
-        self::addHotelAdGroupAd($googleAdsClient, $customerId, $adGroupResourceName);
+            self::addAdGroup($googleAdsClient, $customerId, $campaignResourceName);
+        // Creates an ad group ad.
+        self::addAdGroupAd($googleAdsClient, $customerId, $adGroupResourceName);
     }
 
     /**
-     * Creates a new campaign budget in the specified client account.
+     * Creates a new campaign budget in the specified customer account.
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
@@ -169,7 +158,8 @@ class AddHotelAd
             'delivery_method' => BudgetDeliveryMethod::STANDARD,
             // Sets the amount of budget.
             'amount_micros' => 50000000,
-            // Makes the budget explicitly shared.
+            // Makes the budget explicitly shared. You cannot set it to `false` for Things to do
+            // campaigns.
             'explicitly_shared' => true
         ]);
 
@@ -196,49 +186,45 @@ class AddHotelAd
     }
 
     /**
-     * Creates a new hotel campaign in the specified client account.
+     * Creates a new Things to do campaign in the specified customer account.
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
      * @param string $budgetResourceName the resource name of budget for a new campaign
-     * @param int $hotelCenterAccountId the Hotel Center account ID
-     * @param int $cpcBidCeilingMicroAmount the CPC bid ceiling micro amount
+     * @param int $thingsToDoCenterAccountId the Things to Do Center account ID
      * @return string the resource name of the newly created campaign
      */
-    // [START add_hotel_ad]
-    private static function addHotelCampaign(
+    // [START add_things_to_do_ad]
+    private static function addThingsToDoCampaign(
         GoogleAdsClient $googleAdsClient,
         int $customerId,
         string $budgetResourceName,
-        int $hotelCenterAccountId,
-        int $cpcBidCeilingMicroAmount
+        int $thingsToDoCenterAccountId
     ) {
-        // [START add_hotel_ad_1]
+        // [START add_things_to_do_ad_1]
         // Creates a campaign.
         $campaign = new Campaign([
             'name' => 'Interplanetary Cruise Campaign #' . Helper::getPrintableDatetime(),
-            // Configures settings related to hotel campaigns including advertising channel type
-            // and hotel setting info.
-            'advertising_channel_type' => AdvertisingChannelType::HOTEL,
-            'hotel_setting' => new HotelSettingInfo(['hotel_center_id' => $hotelCenterAccountId]),
+            // Configures settings related to Things to do campaigns including advertising channel
+            // type, advertising channel sub type and travel campaign settings.
+            'advertising_channel_type' => AdvertisingChannelType::TRAVEL,
+            'advertising_channel_sub_type' => AdvertisingChannelSubType::TRAVEL_ACTIVITIES,
+            'travel_campaign_settings'
+                => new TravelCampaignSettings(['travel_account_id' => $thingsToDoCenterAccountId]),
             // Recommendation: Set the campaign to PAUSED when creating it to prevent
             // the ads from immediately serving. Set to ENABLED once you've added
             // targeting and the ads are ready to serve.
             'status' => CampaignStatus::PAUSED,
-            // Sets the bidding strategy to PercentCpc. Only Manual CPC and Percent CPC can be used
-            // for hotel campaigns.
-            'percent_cpc' => new PercentCpc([
-                'cpc_bid_ceiling_micros' => $cpcBidCeilingMicroAmount
-            ]),
+            // Sets the bidding strategy to MaximizeConversionValue. Only this type can be used
+            // for Things to do campaigns.
+            'maximize_conversion_value' => new MaximizeConversionValue(),
             // Sets the budget.
             'campaign_budget' => $budgetResourceName,
             // Configures the campaign network options. Only Google Search is allowed for
-            // hotel campaigns.
-            'network_settings' => new NetworkSettings([
-                'target_google_search' => true,
-            ]),
+            // Things to do campaigns.
+            'network_settings' => new NetworkSettings(['target_google_search' => true])
         ]);
-        // [END add_hotel_ad_1]
+        // [END add_things_to_do_ad_1]
 
         // Creates a campaign operation.
         $campaignOperation = new CampaignOperation();
@@ -251,17 +237,17 @@ class AddHotelAd
         /** @var Campaign $addedCampaign */
         $addedCampaign = $response->getResults()[0];
         printf(
-            "Added a hotel campaign with resource name '%s'.%s",
+            "Added a Things to do campaign with resource name '%s'.%s",
             $addedCampaign->getResourceName(),
             PHP_EOL
         );
 
         return $addedCampaign->getResourceName();
     }
-    // [END add_hotel_ad]
+    // [END add_things_to_do_ad]
 
     /**
-     * Creates a new hotel ad group in the specified campaign.
+     * Creates a new ad group in the specified Things to do campaign.
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
@@ -269,8 +255,8 @@ class AddHotelAd
      *     belong to
      * @return string the resource name of the newly created ad group
      */
-    // [START add_hotel_ad_2]
-    private static function addHotelAdGroup(
+    // [START add_things_to_do_ad_2]
+    private static function addAdGroup(
         GoogleAdsClient $googleAdsClient,
         int $customerId,
         string $campaignResourceName
@@ -280,10 +266,8 @@ class AddHotelAd
             'name' => 'Earth to Mars Cruise #' . Helper::getPrintableDatetime(),
             // Sets the campaign.
             'campaign' => $campaignResourceName,
-            // Sets the ad group type to HOTEL_ADS.
-            // This cannot be set to other types.
-            'type' => AdGroupType::HOTEL_ADS,
-            'cpc_bid_micros' => 10000000,
+            // Sets the ad group type to TRAVEL_ADS. This cannot be set to other types.
+            'type' => AdGroupType::TRAVEL_ADS,
             'status' => AdGroupStatus::ENABLED,
         ]);
 
@@ -298,40 +282,34 @@ class AddHotelAd
         /** @var AdGroup $addedAdGroup */
         $addedAdGroup = $response->getResults()[0];
         printf(
-            "Added a hotel ad group with resource name '%s'.%s",
+            "Added an ad group with resource name '%s'.%s",
             $addedAdGroup->getResourceName(),
             PHP_EOL
         );
 
         return $addedAdGroup->getResourceName();
     }
-    // [END add_hotel_ad_2]
+    // [END add_things_to_do_ad_2]
 
     /**
-     * Creates a new hotel ad group ad in the specified ad group.
+     * Creates a new ad group ad in the specified ad group.
      *
      * @param GoogleAdsClient $googleAdsClient the Google Ads API client
      * @param int $customerId the customer ID
      * @param string $adGroupResourceName the resource name of ad group that a new ad group ad will
      *     belong to
      */
-    // [START add_hotel_ad_3]
-    private static function addHotelAdGroupAd(
+    // [START add_things_to_do_ad_3]
+    private static function addAdGroupAd(
         GoogleAdsClient $googleAdsClient,
         int $customerId,
         string $adGroupResourceName
     ) {
-        // Creates a new hotel ad.
-        $ad = new Ad([
-            'hotel_ad' => new HotelAdInfo(),
-        ]);
-
-        // Creates a new ad group ad and sets the hotel ad to it.
+        // Creates a new ad group ad and sets a travel ad info.
         $adGroupAd = new AdGroupAd([
-            'ad' => $ad,
-            // Set the ad group ad to enabled.  Setting this to paused will cause an error
-            // for hotel campaigns.  For hotels pausing should happen at either the ad group or
-            // campaign level.
+            'ad' => new Ad(['travel_ad' => new TravelAdInfo()]),
+            // Set the ad group ad to enabled. Setting this to paused will cause an error for Things
+            // to do campaigns. Pausing should happen at either the ad group or campaign level.
             'status' => AdGroupAdStatus::ENABLED,
             // Sets the ad group.
             'ad_group' => $adGroupResourceName
@@ -348,12 +326,12 @@ class AddHotelAd
         /** @var AdGroupAd $addedAdGroupAd */
         $addedAdGroupAd = $response->getResults()[0];
         printf(
-            "Added a hotel ad group ad with resource name '%s'.%s",
+            "Added an ad group ad with resource name '%s'.%s",
             $addedAdGroupAd->getResourceName(),
             PHP_EOL
         );
     }
-    // [END add_hotel_ad_3]
+    // [END add_things_to_do_ad_3]
 }
 
-AddHotelAd::main();
+AddThingsToDoAd::main();
